major file refactoring
