+++ ToolOwnershipTracker/ToolOwnershipTracker/tests/unit tests/userTests.py
@@ -1,0 +1,6 @@
+from ToolOwnerShipTracker.models import user
+from ToolOwnerShipTracker.classes.Users import Users
+from django.test import TestCase
+import unittest
+import django
+django.setup()
